backend/requirements.txt
flask
flask-cors
python-dotenv
groq
pyautogui
Pillow
requests
pydantic
pyperclip